Add support for arrays of types
If type is an array, we currently expect that one of the values in the
array is null, and the other is either "object" or "array".

It would certainly be possible to expand support for more intricate type
defs, but I'd suggest that we wait until we encounter those cases before
supporting them.

See discussion above func primaryType in defaults.go